Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
google-generativeai==0.8.0
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.0
markdown==3.5.1
easyocr==1.7.1
PyPDF2==3.0.1
python-docx==1.1.0